requirements.txt
gitpython==3.1.0
prometheus_client==0.8.0
pandas==1.2.1